#8 - reorganizing code before implementation of exploit protection.
