+++ src/main/java/com/marcuslull/springauthserver/model/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.marcuslull.springauthserver;
+package com.marcuslull.springauthserver.model;
 
 public record LoginRequest(String userName, String password) {
 }
